fix 32bit/64bit regressor file incompatibility

+++ vowpalwabbit/global_data.h
@@ -217,7 +217,7 @@ struct vw {
   size_t normalized_idx; //offset idx where the norm is stored (1 or 2 depending on whether adaptive is true)
   size_t feature_mask_idx; //offset idx where mask is stored
 
-  size_t lda;
+  uint32_t lda;
   float lda_alpha;
   float lda_rho;
   float lda_D;
